lambda/pyproject.toml
[project]
name = "ekim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-lambda-powertools>=3.18.0",
    "boto3>=1.40.2",
    "dnspython>=2.7.0",
    "fastapi>=0.116.1",
    "mangum>=0.19.0",
    "types-boto3[full]>=1.40.2",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.115",
]
